New/services/reporting-agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
jinja2==3.1.3
PyGithub==2.2.0
slack-sdk==3.26.2
